move static functions outside of class definition (#20)
All static or static-eligible methods (methods that don't reference the
`self` argument) have been moved out of the checker classes and up to
the module level. This enhances the functions' testability and makes the
checker class code easier to read as they now only contain visitor
methods or methods that compute message criteria and add messages.